+++ bobtemplates/plone/addon.py
@@ -3,6 +3,7 @@
 from bobtemplates.plone.base import git_commit
 from bobtemplates.plone.base import git_init
 from bobtemplates.plone.base import make_path
+from bobtemplates.plone.utils import run_isort
 
 import os
 import shutil
@@ -137,6 +138,7 @@ def pre_ask(configurator):
 
 def post_render(configurator):
     _cleanup_package(configurator)
+    run_isort(configurator)
     git_init_status = git_init(configurator)
     if git_init_status:
         git_commit(